[Refactor] 제작 UI class 분리
