+++ Assets/SHG/Scripts/UI/RecipeView.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 46ddef18bd1d992fdb338bb2b68b9886
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
