+++ Assets/SHG/Scripts/UI/RecipeWindow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b39c4cea5e66e5e3887a7b2ae2b6a1ae
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
